added get checklist by id

+++ dilpoma-backend/src/main/java/com/edu/netc/bakensweets/model/Checklist.java
@@ -11,5 +11,5 @@ public class Checklist {
     private long id;
     private String name;
     private long deviceId;
-    private int accountId;
+    private long accountId;
 }
